% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbst-class.R
\docType{methods}
\name{affine,scidbst-method}
\alias{affine,scidbst-method}
\title{Returns the affine transformation}
\usage{
\S4method{affine}{scidbst}(x)
}
\arguments{
\item{x}{scidbst object}
}
\value{
a numeric matrix containing the affine transformation parameter
}
\description{
The function returns the stored affine transformation. The matrix has a dimensionality of 2x3 and contains the following values:
x0,xres(x),xshear(x) \\ y0, yshear(x), yres(x)
}

