% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine.R
\docType{methods}
\name{affine,scidbst}
\alias{affine}
\alias{affine,scidbst}
\alias{affine<-,scidb}
\alias{affine<-,scidbst}
\title{Handling for the affine transformation on scidb(st) arrays}
\usage{
\S4method{affine}{scidbst}(x)

\S4method{affine}{scidbst,matrix}(x) <- value

\S4method{affine}{scidb,matrix}(x) <- value
}
\arguments{
\item{x}{scidbst object}
}
\value{
a numeric matrix containing the affine transformation parameter
}
\description{
Returns the affine transformation as a 2x3 matrix (x0,xres(x),xshear(x) \\ y0, yshear(x), yres(x)) or sets the affine transformation
for a scidb or scidbst object.
}
\examples{
\dontrun{
 # Getter
 .scidbst = scidbst("ref_array")
 aff.trans = affine(.scidbst)

 # Setter
 .scidb = scidb("an_array")
 m = matrix(c(1000,1,0, 1000,0,1),byrow=TRUE,nc=3,nr=2)
 affine(.scidb) <- m

}
}

