% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\docType{methods}
\name{aggregate,scidbst-method}
\alias{aggregate,scidbst-method}
\title{Aggregates a 'scidbst' object over the given dimensions and/or attributes}
\usage{
aggregate(x, by, FUN, window, variable_window)
}
\arguments{
\item{x}{A \code{scidbst} object.}

\item{by}{optional single character string or a list of array dimension and/or attribute names to group by;
or a \code{scidb} object to group by. Not required for \code{windowed} and grand aggregates.}

\item{FUN}{a character string representing a SciDB aggregation expression or a reduction function.}

\item{window}{optional, if specified, perform a moving window aggregate along the specified coordinate windows.}

\item{variable_window}{optional, if specified, perform a moving window aggregate over successive data values along the
coordinate dimension axis specified by \code{by}.}
}
\description{
Due to the fact that scidbst arrays can have dimensions that are referenced in space and/or time, this function
uses 'scidb's aggregate function for performing the aggregation itself and it manages the metadata information
about the references / extents and alike.
}
\details{
The scidbst package also provides functions to directly aggregate over space or time without the need to
specify the dimensions that need to be aggregated by. For more information see: \code{\link{aggregate.t}} and
\code{\link{aggregate.sp}}
}
\examples{
scidbconnect(...)
scidbst.obj = scidbst(array_name) #array with spatial and temporal dimension (x,y,t)
agg.1 = aggregate(x=scidbst.obj,by=list("y","x"),FUN="avg(band1)") # aggreagtes over time, result something similar to a raster
agg.2 = aggregate(x=scidbst.obj,by=list("t"),FUN="avg(band1)") # aggregate over space, result something like a time series
agg.3 = aggregate(x=scidbst.obj,by=list("y","x","t"),FUN=count) # count cells that are aggregated = total number of cells over all dimensions
}
\seealso{
\code{\link{aggregate,scidb-method}}, \code{\link{aggregate.t}}, \code{\link{aggregate.sp}}
}

