% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\docType{methods}
\name{aggregate.sp,scidbst-method}
\alias{aggregate.sp}
\alias{aggregate.sp,scidbst-method}
\title{Aggregate over space}
\usage{
aggregate(x, by, FUN, window, variable_window)
}
\arguments{
\item{x}{A \code{scidbst} object.}

\item{by}{optional single character string or a list of array dimension and/or attribute names to group by additionally to
the spatial dimensions}

\item{FUN}{a character string representing a SciDB aggregation expression or a reduction function.}

\item{window}{optional, if specified, perform a moving window aggregate along the specified coordinate windows.}

\item{variable_window}{optional, if specified, perform a moving window aggregate over successive data values along the
coordinate dimension axis specified by \code{by}.}
}
\value{
scidbst object with the spatial resolution extended to the whole image extent (spatial dimension is removed), often times
something similar to a time series will remain
}
\description{
This function aggregates over space leaving a scidbst array without spatial dimensions. The spatial information will remain
on the R object. The spatial resolution will be increased to the whole spatial dimensions (one cell captures the whole image). And
the spatial extent will remain the same.
}
\note{
the aggregate scidbst object will loose its spatial reference after evaluation, due to the fact that the spatial
dimensions are removed
}
\examples{
\dontrun{
scidbconnect(...)
scidbst.obj = scidbst(array_name)
aggt = aggregate.sp(scidbst.obj,FUN="avg(attribute1)") # returns something similar to a timeseries with aggregated values over the spatial dimension
}
}

