% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\docType{methods}
\name{aggregate.t,scidbst-method}
\alias{aggregate.t,scidbst-method}
\title{aggregates over time}
\usage{
\S4method{aggregate.t}{scidbst}(x, ...)
}
\value{
scidbst array with a temporal resolution of the whole time series
}
\description{
This function aggregates a scidbst array for the dimension time. This means the resulting array will stripped from
the temporal dimension.
}

