% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\docType{methods}
\name{apply.fun,scidbst,function-method}
\alias{apply.fun,scidbst,function-method}
\title{Applys a custom function on chunks of an array}
\usage{
\S4method{apply.fun}{scidbst,`function`}(x, f, array, packages,
  parallel = FALSE, cores = 1, aggregates, output, logfile, dim, dim.spec,
  method = "rexec", ...)
}
\arguments{
\item{x}{scidbst array}

\item{f}{function}
}
\value{
scidbst array
}
\description{
This function applies a custom function on a scidbst array chunk using r_exec.
}
\details{
The function that can be stated has the following description "function(x,...) {}". The x parameter is a
data.frame of the attributes stored in one chunk. In most cases you are advised to transform the array to have the
dimension values as attributes if you need those to perform calculations. The function will be passed on to the \link[plyr]{ddply}
function.
}

