% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{as-data.frame}
\alias{as-data.frame}
\title{Creates a data.frame from the scidbst object}
\description{
Creates a data.frame from the scidbst object
}
\seealso{
Other scidbst: \code{\link{as-RasterBrick}},
  \code{\link{as-STSDF}},
  \code{\link{as-SpatialPointsDataFrame}},
  \code{\link{as-xts}}
}

