% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{as-RasterBrick}
\alias{as-RasterBrick}
\title{Creates a RasterBrick from scidbst}
\description{
Creates a RasterBrick from scidbst
}
\seealso{
Other scidbst: \code{\link{as-STSDF}},
  \code{\link{as-SpatialPointsDataFrame}},
  \code{\link{as-data.frame}}, \code{\link{as-scidb}},
  \code{\link{as-xts}}
}

