% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{as-scidb}
\alias{as-scidb}
\title{Returns the scidb proxy part of the scidbst object}
\description{
Returns the scidb proxy part of the scidbst object
}
\seealso{
Other scidbst: \code{\link{as-RasterBrick}},
  \code{\link{as-STSDF}},
  \code{\link{as-SpatialPointsDataFrame}},
  \code{\link{as-data.frame}}, \code{\link{as-xts}}
}

