% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{as-scidbst,data.frame}
\alias{as-scidbst,data.frame}
\title{As data.frame}
\description{
Coerces the scidbst object into a data.frame object. The dimension values will NOT be recalculated!
}
\note{
Dimensions and attributes will be treated as columns of the data.frame in the same way as \code{\link{[.scidb}}
does.
}
\seealso{
\link{[.scidb}

Other coerce.scidbst: \code{\link{as-scidbst,RasterBrick}},
  \code{\link{as-scidbst,STSDF}},
  \code{\link{as-scidbst,SpatialPointsDataFrame}},
  \code{\link{as-scidbst,scidb}},
  \code{\link{as-scidbst,xts}}
}

