% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{as-scidbst,RasterBrick}
\alias{as-scidbst,RasterBrick}
\title{As RasterBrick}
\description{
Coerces a scidbst object into a \code{RasterBrick} object. The information of the array will be downloaded to the local client
and will be hold in memory. The spatial dimension indices will be recalculated into spatial coordinates during the process.
}
\seealso{
Other coerce.scidbst: \code{\link{as-scidbst,STSDF}},
  \code{\link{as-scidbst,SpatialPointsDataFrame}},
  \code{\link{as-scidbst,data.frame}},
  \code{\link{as-scidbst,scidb}},
  \code{\link{as-scidbst,xts}}
}

