% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{as-scidbst,scidb}
\alias{as-scidbst,scidb}
\title{Returns the scidb proxy part of the scidbst object}
\description{
Returns the scidb proxy part of the scidbst object
}
\seealso{
Other coerce.scidbst: \code{\link{as-scidbst,RasterBrick}},
  \code{\link{as-scidbst,STSDF}},
  \code{\link{as-scidbst,SpatialPointsDataFrame}},
  \code{\link{as-scidbst,data.frame}},
  \code{\link{as-scidbst,xts}}
}

