% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{as-scidbst,SpatialPointsDataFrame}
\alias{as-scidbst,SpatialPointsDataFrame}
\title{As SpatialPointsDataFrame}
\description{
Coerces a scidbst object into a \code{\link{SpatialPointsDataFrame}} by downloading the array information and recalculating
the spatial dimension indices into spatial coordinates.
}
\seealso{
Other coerce.scidbst: \code{\link{as-scidbst,RasterBrick}},
  \code{\link{as-scidbst,STSDF}},
  \code{\link{as-scidbst,data.frame}},
  \code{\link{as-scidbst,scidb}},
  \code{\link{as-scidbst,xts}}
}

