% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{as-scidbst,STSDF}
\alias{as-scidbst,STSDF}
\title{As STSDF}
\description{
Coerces a scidbst object into a STSDF object. Spatial and temporal dimension indices are calculated into real-world
coordinates using existing spatio-temporal references.
}
\seealso{
Other coerce.scidbst: \code{\link{as-scidbst,RasterBrick}},
  \code{\link{as-scidbst,SpatialPointsDataFrame}},
  \code{\link{as-scidbst,data.frame}},
  \code{\link{as-scidbst,scidb}},
  \code{\link{as-scidbst,xts}}
}

