% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{as-scidbst,xts}
\alias{as-scidbst,xts}
\title{As xts}
\description{
Coerces a scidbst object into a xts time series.
}
\seealso{
Other coerce.scidbst: \code{\link{as-scidbst,RasterBrick}},
  \code{\link{as-scidbst,STSDF}},
  \code{\link{as-scidbst,SpatialPointsDataFrame}},
  \code{\link{as-scidbst,data.frame}},
  \code{\link{as-scidbst,scidb}}
}

