% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{as-SpatialPointsDataFrame}
\alias{as-SpatialPointsDataFrame}
\title{Creates a SpatialPointsDataFrame from scidbst}
\description{
Creates a SpatialPointsDataFrame from scidbst
}
\seealso{
Other scidbst: \code{\link{as-RasterBrick}},
  \code{\link{as-STSDF}}, \code{\link{as-data.frame}},
  \code{\link{as-xts}}
}

