% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{as-STSDF}
\alias{as-STSDF}
\title{Creates a STSDF from scidbst}
\description{
Creates a STSDF from scidbst
}
\seealso{
Other scidbst: \code{\link{as-RasterBrick}},
  \code{\link{as-SpatialPointsDataFrame}},
  \code{\link{as-data.frame}}, \code{\link{as-xts}}
}

