% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{as}
\alias{as}
\title{Creates a SpatialPointsDataFrame from scidbst}
\description{
Creates a SpatialPointsDataFrame from scidbst

Creates a RasterBrick from scidbst

Creates a STSDF from scidbst

Creates a RasterBrick from scidbst
}

