% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRS-class.R
\docType{methods}
\name{copySRS,scidbst,scidbst-method}
\alias{copySRS,scidb,scidbst-method}
\alias{copySRS,scidbst,scidbst-method}
\title{Copy spatial reference}
\usage{
\S4method{copySRS}{scidbst,scidbst}(x, y)

\S4method{copySRS}{scidb,scidbst}(x, y)
}
\arguments{
\item{x}{scidbst or scidb object}

\item{y}{scidbst object}
}
\value{
modified x
}
\description{
Copies the spatial reference systems from scidbst object y to scidb(st) object x.
}

