% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRS-class.R
\docType{methods}
\name{copyTRS,scidbst,scidbst-method}
\alias{copyTRS,scidb,scidbst-method}
\alias{copyTRS,scidbst,scidbst-method}
\title{Copy temporal reference}
\usage{
\S4method{copyTRS}{scidbst,scidbst}(x, y)

\S4method{copyTRS}{scidb,scidbst}(x, y)
}
\arguments{
\item{x}{scidbst or scidb object}

\item{y}{scidbst object}
}
\value{
modified x
}
\description{
Copies the temporal reference systems from scidbst object y to scidbst object x.
}

