% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\docType{methods}
\name{crop,scidbst-method}
\alias{crop,scidbst-method}
\title{crop function}
\usage{
\S4method{crop}{scidbst}(x, y, snap = "near", ...)
}
\arguments{
\item{x}{scidbst object}

\item{y}{Extent object, or any object from which an Extent object can be extracted}

\item{snap}{Character. One of 'near', 'in', or 'out', for use with alignExtent}

\item{...}{Additional arguments as for writeRaster}
}
\value{
scidbst object with refined spatial extent
}
\description{
This function creates a spatial subset of a scidbst array and returns the subset scidbst object.
}

