% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRS-class.R
\docType{methods}
\name{crs,scidbst-method}
\alias{crs,scidbst-method}
\title{Returns the Coordinate Reference System}
\usage{
\S4method{crs}{scidbst}(x, ...)
}
\arguments{
\item{x}{scidbst object}
}
\value{
\link[rgdal]{CRS}
}
\description{
Returns the coordinate reference system of a scidbst object
}

