% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRS-class.R
\docType{methods}
\name{crs,scidbst-method}
\alias{crs,scidbst-method}
\alias{crs<-,ANY,CRS-method}
\alias{crs<-,scidbst,CRS-method}
\title{Returns the Coordinate Reference System}
\usage{
\S4method{crs}{scidbst}(x, ...)

\S4method{crs}{scidbst,CRS}(x) <- value

\S4method{crs}{ANY,CRS}(x) <- value
}
\arguments{
\item{x}{scidbst object}
}
\value{
\link[rgdal]{CRS}
}
\description{
Returns the coordinate reference system of a scidbst object in form of a CRS object.
}
\note{
The changes only apply in R. To persist the changes make the object truly spatial by setting also the
affine transformation as well as the spatial extent and then run scidbsteval.
}

