% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulate.R
\docType{methods}
\name{cumulate,scidbst-method}
\alias{cumulate,scidbst-method}
\alias{cumulate.scidbst}
\title{Cumulate function for scidbst}
\usage{
\S4method{cumulate}{scidbst}(x, expression, dimension)
}
\arguments{
\item{x}{scidbst object}

\item{expression}{any valid SciDB aggregate expression as a character string}

\item{dimension}{optional parameter, the dimension along which the cumulative sum or product will be calculates}
}
\value{
a scidbst object
}
\description{
This function wraps scidbs cumulate function in order to maintain the scidbst object with the dimension references.
}
\seealso{
\link{scidb::cumulate}
}

