% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulate.R
\docType{methods}
\name{cumulate,scidbst}
\alias{cumulate,scidbst}
\title{Cumulate function for scidbst objects}
\usage{
\S4method{cumulate}{scidbst}(x, expression, dimension)
}
\arguments{
\item{x}{scidbst object}

\item{expression}{any valid SciDB aggregate expression as a character string}

\item{dimension}{optional parameter, the dimension along which the cumulative sum or product will be calculates}
}
\value{
a scidbst object
}
\description{
This function wraps scidbs cumulate function in order to maintain the scidbst object with the dimension references.
}
\note{
Please use scidb::cumulate, when you operate on pure scidb objects, since package 'scidb' does not export this function
as a S4 method.
}
\examples{
\dontrun{
 scidbconnect(...)
 chicago = scidbst("chicago_sts")
 cumsum.band1 = cumulate.scidbst(chicago,"sum(band1)","t")
}

}
\seealso{
\link{scidb::cumulate}
}

