% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{datetime,scidbst-method}
\alias{datetime,scidbst-method}
\title{Calculates a POSIXt from a time index}
\usage{
\S4method{datetime}{scidbst}(x, n)
}
\arguments{
\item{x}{scidbst object}

\item{n}{integer value}
}
\value{
POSIXt object for the time index n
}
\description{
The function calculates the appropriate POSIXt object given an scidbst array and a time index.
}

