% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension.R
\docType{methods}
\name{dimensions,scidbst}
\alias{dimensions,scidb}
\alias{dimensions,scidbst}
\alias{tdim,TRS}
\alias{tdim,scidbst}
\alias{xdim,SRS}
\alias{xdim,scidbst}
\alias{ydim,SRS}
\alias{ydim,scidbst}
\title{Dimension operations}
\usage{
\S4method{dimensions}{scidbst}(obj)

\S4method{dimensions}{scidb}(obj)

\S4method{xdim}{scidbst}(x)

\S4method{xdim}{SRS}(x)

\S4method{ydim}{scidbst}(x)

\S4method{ydim}{SRS}(x)

\S4method{tdim}{TRS}(x)

\S4method{tdim}{scidbst}(x)
}
\arguments{
\item{obj}{scidbst object}

\item{x}{scidbst object}
}
\value{
character refering to the name of the selected dimension(s).
}
\description{
The listed functions are used to find the correct naming for the dimensions in order to address them correctly in other operations.
The x dimension relates to a West-East dimension, whereas the y dimension relates to the North-South dimension. Also the function tdim
returns the name of the temporal dimension used in SciDB.
}
\examples{
\dontrun{
starr = scidbst("st_array")
dimensions(starr)
xdim(starr)
ydim(starr)
tdim(starr)
}
}
\seealso{
\code{\link[scidb]{dimensions}}
}

