% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\docType{methods}
\name{equalize,scidbst,scidbst-method}
\alias{equalize,scidbst,scidbst-method}
\title{Equalizes the dimensionality of two scidbst arrays}
\usage{
\S4method{equalize}{scidbst,scidbst}(x, y, type, storeTemp, saf, taf)
}
\arguments{
\item{x}{source scidbst array}

\item{y}{target scidbst array}

\item{type}{character one of c("S","T","ST") two equalize spatial, temporal or spatio-temporal}

\item{storeTemp}{logical - whether or not to preevaluate the arrays and store them temporarily}

\item{taf}{the temporal aggregation function to apply}

\item{raf}{the spatial aggregation function to apply}
}
\value{
the modified scidbst array source array with the dimensional representation as the target
}
\description{
This function modifies the dimension properties of a source scidbst array to match the spatial
and/or temporal resolution of a target array. Therefore scidb will perform either an aggregation
for upscaling (higher to lower resolution) or it breaks down cells for a downscaling (lower to higher
resolution)
}
\seealso{
\href{http://www.paradigm4.com/HTMLmanual/15.7/scidb_ug/Aggregates.html}{scidb aggregation functions}
}

