% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{estimateFileSize,scidbst-method}
\alias{estimateFileSize,scidbst-method}
\title{Estimates the file size for a scidbst array}
\usage{
\S4method{estimateFileSize}{scidbst}(x, unit = "MB")
}
\arguments{
\item{x}{scidbst array}

\item{unit}{(optional) one of c("KB","MB","GB") to specify the digital storage unit. Default is MB}
}
\value{
numeric the estimated file size in the unit specified
}
\description{
This function uses the meta information on a scidbst array to estimate the potential file size.
}

