% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent.R
\docType{methods}
\name{extent,scidbst}
\alias{extent,scidbst}
\alias{xmax,scidbst}
\alias{xmin,scidbst}
\alias{ymax,scidbst}
\alias{ymin,scidbst}
\title{Get the spatial extent of a scidbst array}
\usage{
\S4method{extent}{scidbst}(x)

\S4method{xmin}{scidbst}(x)

\S4method{ymin}{scidbst}(x)

\S4method{xmax}{scidbst}(x)

\S4method{ymax}{scidbst}(x)
}
\arguments{
\item{x}{scidbst array}
}
\value{
Extent-class object
}
\description{
This function let you retrieve the spatial extent as a Extent-class object. With the additional functions xmin, xmax, ymin and ymax
particular information about the extent can be retrieved directly from the scidbst object.
}
\examples{
\dontrun{
arr = scidbst("sp_ref_array")
extent(arr)
xmin(arr)
xmax(arr)
ymin(arr)
ymax(arr)
}
}

