% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getterSetter.R
\docType{methods}
\name{getLayer,scidbst-method}
\alias{getLayer,scidbst-method}
\title{Get Layer}
\usage{
\S4method{getLayer}{scidbst}(x, layer, ...)
}
\arguments{
\item{x}{scidbst object with raster data}

\item{layer}{the name or index of the attribute}

\item{...}{additional parameters that are passed to \link{raster::subset}}
}
\value{
A RasterLayer object
}
\description{
If the data is loaded as a raster brick, this function will replace the raster subset function, since scidb
uses a function with the same name to subset directly from the scidb database.
}
\seealso{
\code{\link{raster::subset}}
}

