% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRS-class.R
\docType{methods}
\name{getRefPeriod}
\alias{getRefPeriod}
\alias{getRefPeriod,TRS-method}
\title{Creates a character expression for the reference time period}
\usage{
\S4method{getRefPeriod}{scidbst}(x)

\S4method{getRefPeriod}{TRS}(x)
}
\arguments{
\item{x}{scidbst object}
}
\value{
a character describing the reference period for a time unit
}
\description{
Returns the reference period of a scidbst object, e.g. P1D, P16D or P1M. It combines the temporal resolution with an abbreviation
for the time unit. Usually the expression is used as "dt" (delta t).
}

