% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getterSetter.R
\docType{methods}
\name{getTDim,scidbst-method}
\alias{getTDim,scidbst-method}
\title{Get temporal dimension name}
\usage{
\S4method{getTDim}{scidbst}(x)
}
\arguments{
\item{x}{scidbst object}
}
\value{
The temporal dimension name or the first dimension name
}
\description{
Returns the name of the temporal dimension of a scidbst array. If the object is not temporal it will
return the first dimension name.
}

