% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValues.R
\docType{methods}
\name{getValues,scidbst,missing,missing-method}
\alias{getValues,scidbst,missing,missing-method}
\title{getValues method}
\usage{
\S4method{getValues}{scidbst,missing,missing}(x)
}
\arguments{
\item{x}{scidbst object}
}
\value{
vector or matrix of raster values
}
\description{
This function retrieves data from the remote scidb database and stores them internally like a Raster* object. This function
work in principle like scidbs array materialization 'array []'. However this function also needs the multidimensional array
to be reduced to a simple 2 dimensional array (spatial dimensions)
}
\examples{
\dontrun{
scidbconnect(...)
array_proxy = scidbst(...)
getValues(array_proxy)
}
}

