% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getterSetter.R
\docType{methods}
\name{getYDim,scidbst-method}
\alias{getYDim,scidbst-method}
\title{Get spatial y dimension name}
\usage{
\S4method{getYDim}{scidbst}(x)
}
\arguments{
\item{x}{scidbst object}
}
\value{
The name of the spatia North-South dimension or the first dimension name
}
\description{
Returns the name of the spatial dimension of a scidbst array, that relates to a North-South dimension. If the object is not
spatially referenced it will return the first dimension name.
}

