% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRS-class.R
\docType{methods}
\name{is.spatial,scidbst-method}
\alias{is.spatial,scidbst-method}
\title{Check for a scidbst object having a spatial reference}
\usage{
\S4method{is.spatial}{scidbst}(x)
}
\arguments{
\item{x}{scidbst object}
}
\value{
logical whether or not the object has a spatial reference
}
\description{
The function checks for certain parameter considered to describe the spatial reference.
}

