% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRS-class.R
\docType{methods}
\name{is.temporal,scidbst-method}
\alias{is.temporal,scidbst-method}
\title{Check for a scidbst object having a temporal reference}
\usage{
\S4method{is.temporal}{scidbst}(x)
}
\arguments{
\item{x}{scidbst object}
}
\value{
logical whether or not the object has a temporal reference
}
\description{
The function checks for certain parameter considered to describe the temporal reference.
}

