% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\docType{methods}
\name{join,scidbst,scidbst-method}
\alias{join,scidbst,scidbst-method}
\title{Joins the attributes of x and y}
\usage{
\S4method{join}{scidbst,scidbst}(x, y, storeTemp = FALSE)
}
\arguments{
\item{x}{scidbst object}

\item{y}{scidbst object}

\item{storeTemp}{logical whether or not some arrays are stored temporarily during the operation}
}
\value{
scidbst object with a combined object
}
\description{
This function executes a join operation on two scidbst arrays, where all attributes of x and y are joined. To make it more concrete
the array dimension structure of y will be used as target structure for the dimensions. Read the method as "join x into y".
}
\details{
The join function will assume that the two arrays are similar in their dimensional representation. To be more concrete, x will
be modified into the array structure including lower and upper dimensional boundaries, overlaps and chunksizes. To make sure the arrays
are similar start by using \code{\link{equalize}} to smooth the transition by equalizing the cell resolution.
}
\seealso{
\code{\link[scidb]{merge.scidb}}, \code{\link{equalize}}
}

