% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\docType{methods}
\name{join,scidbst}
\alias{join,scidbst}
\title{Joins the attributes of x and y}
\usage{
\S4method{join}{scidbst,scidbst}(x, y, storeTemp = FALSE, name, raf = "avg")
}
\arguments{
\item{x}{scidbst object}

\item{y}{scidbst object}

\item{storeTemp}{logical whether or not some arrays are stored temporarily during the operation}

\item{name}{The name of the joined output array (required if storeTemp==TRUE)}

\item{raf}{the aggregation function to be applied during the spatial regrid}
}
\value{
scidbst object with a combined object
}
\description{
This function executes a join operation on two scidbst arrays, where all attributes of x and y are joined.
}
\details{
At this point the join operation works spatial arrays only (this will be extended in the future). As a merge strategy the array with the
lower resolution will be choosen as the target array. The array with the higher resolution will be regridded into the target array structure, so that
cells can be merged by transfering the coordinates into the target arrays dimension indices.
}

