% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbst-class.R
\docType{methods}
\name{names,scidbst-method}
\alias{names,scidbst-method}
\title{names function}
\usage{
\S4method{names}{scidbst}(x)
}
\arguments{
\item{x}{scibst object}
}
\value{
vector of character containing the names of dimensions and attributes
}
\description{
Returns the names of the dimensions and attributes used in the remote scidb array.
}
\note{
This function overwrites the standard S3 names function of raster.
}

