% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncol.R
\docType{methods}
\name{ncol,scidbst-method}
\alias{ncol,scidbst-method}
\title{Number of columns}
\usage{
\S4method{ncol}{scidbst}(x)
}
\arguments{
\item{x}{scibst}
}
\value{
number of columns for a scidbst object
}
\description{
returns the number of columns
}

