% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrow.R
\docType{methods}
\name{nrow,scidbst-method}
\alias{nrow,scidbst-method}
\title{Number of rows}
\usage{
\S4method{nrow}{scidbst}(x)
}
\arguments{
\item{x}{scibst}
}
\value{
number of rows for a scidbst object
}
\description{
returns the number of rows
}

