% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelts.R
\name{pixelts,scidbst}
\alias{pixelts,scidbst}
\alias{pixelts}
\title{Pixel time series extraction}
\usage{
pixelts(obj, x)
}
\arguments{
\item{obj}{scidbst object}

\item{x}{coordinates of pixel where the time series will be extracted}
}
\description{
Extracts the time series of a SciDB array at a given coordinate.
}
