% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\docType{methods}
\name{project,scidbst}
\alias{project,scidbst}
\title{Project onto a subset of attributes in a scidbst object}
\usage{
\S4method{project}{scidbst}(x, attributes)
}
\arguments{
\item{x}{\code{scidbst} object}

\item{attributes}{a character vector of array attribute names or a numeric vector of attribute positions}
}
\value{
a new \code{scidbst} object
}
\description{
This function will create a selection of attributes by passing the attribute statement to the scidb project function.
}
\seealso{
\code{\link[scidb]{project}}
}

