% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAll.R
\docType{methods}
\name{readAll,scidbst-method}
\alias{readAll,scidbst-method}
\title{readAll}
\usage{
\S4method{readAll}{scidbst}(object)
}
\arguments{
\item{object}{scidbst object}
}
\value{
the modified scidbst object with values in memory (stored as a RasterBrick)
}
\description{
Like \code{raster::getValues()} this function is called to retrieve the values from the scidb database and store them in memory. It
differs from getValues in the fact that the scidbst object will be manipulated and returned back.
}
\note{
Currently the number of dimensions for this array needs to be 1 or 2, in order to limit the data download traffic.
}
\examples{
\dontrun{
scidbconnect(...)
scidbst.obj = scidbst(array_name) # e.g. a spatio-temporal object with dimension (y,x,t)
sliced = slice(scidbst.obj, "t",0)
spplot(sliced) #note that the dimensionality of scidbst.obj needs to be 2 or less
}
}

