% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repart.R
\docType{methods}
\name{repart,scidbst-method}
\alias{repart,scidbst-method}
\title{Repartion a scidbst array}
\usage{
\S4method{repart}{scidbst}(x, schema, upper, chunk, overlap)
}
\arguments{
\item{x}{the scidbst array}

\item{schema}{The new schema as a string}

\item{upper}{A vector of upper boundaries for the array}

\item{chunk}{A vector of chunksizes for each dimension}

\item{overlap}{A vector of chunk overlaps for the repartitioned array}
}
\value{
scidbst array with modified chunks
}
\description{
Repartions a scidb array with new chunk and overlap sizes. It will move the values into new chunks.
}
\note{
This function might take a while, because all data values need to be moved (restored) accordingly. You might either state
a valid schema parameter or build a schema from the parameters "upper", "chunk" and "overlap".
}

