% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repart.R
\docType{methods}
\name{repart,scidbst}
\alias{repart,scidbst}
\title{Repartion a scidbst array}
\usage{
\S4method{repart}{scidbst}(x, schema, upper, chunk, overlap)
}
\arguments{
\item{x}{the scidbst array}

\item{schema}{The new schema as a string}

\item{upper}{A vector of upper boundaries for the array}

\item{chunk}{A vector of chunksizes for each dimension}

\item{overlap}{A vector of chunk overlaps for the repartitioned array}
}
\value{
scidbst array with modified chunks
}
\description{
Repartions a scidb array with new chunk and overlap sizes. It will move the values into new chunks.
}
\note{
This function might take a while, because all data values need to be moved (restored) accordingly. You might either state
a valid schema parameter or build a schema from the parameters "upper", "chunk" and "overlap".
}
\examples{
\dontrun{
   scidbconnect()
   trmm.sub = scidbst("trmm_sub")

   # repartition the array by stating a new schema with the same attributes and dimensions
   trmm.reparted = repart(trmm.sub,schema="<band1:double,dimy:double,dimx:double,dimt:double>[y=0:399,30,1,x=0:1439,30,1,t=0:*,365,6]")
}
}
\seealso{
\code{\link[scidb]{repart}}
}

