% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrid.R
\docType{methods}
\name{regrid,scidbst-method}
\alias{regrid,scidbst-method}
\alias{resample,scidbst,Raster-method}
\title{Regrid / Resample operator for scidbst objects}
\usage{
\S4method{regrid}{scidbst}(x, grid, expr)

\S4method{resample}{scidbst,Raster}(x, y, expr)
}
\arguments{
\item{x}{The scidbst object}

\item{grid}{a vector of grid sizes having the same length as dimensions(x)}

\item{expr}{(optional) aggregation function applied to every attribute on the grid, or a quoted SciDB aggregation expression}

\item{y}{raster object}
}
\value{
scidbst object with new resolution information (adapted affine transformation)
}
\description{
This function changes the resolution of certain dimensions of the scidbst object. 'Regrid' is in this case the targeted scidb operation and reflects
a more general regrid approach. 'Resample' is based on the respective function of the raster package. The latter will simply perform a regrid on the
spatial dimensions. The grid parameter reflects the number of cells per dimension that are used as a block for resampling. As for the current development
of scidb, the aggregation functions are quite limited (min/max, average, sum, standard deviation) and do not support more elaborated function like bilinear
interpolation.
}
\note{
For information on the aggregation statements in scidb have a look on the supported AFL functions (\url{http://paradigm4.com/HTMLmanual/13.3/scidb_ug/ch12.html})
}
\examples{
\dontrun{
scidbconnect(...)
scibdst.obj = scidbst(array_name)
# 50x50 upscaling with scidb expression as aggregation function
regridded = regrid(scidbst.obj,c(50,50,1),"avg(attribute1)") #resample (upscaling 50x50 dimension units, here dimensions: y,x,time)
# 50x50 upscaling with sum aggregation function
regridded = regrid(scidbst.obj,c(50,50,1),sum) #resample (upscaling 50x50 dimension units, here dimensions: y,x,time)

r = raster(scidbst.obj, nrows=300, ncols=200) # resample by setting the number of rows and columns of the target
resampled = resample(scidbst.obj,r,"avg(attribute1)")
}

}
\seealso{
\code{\link{regrid,scidb-method}} for original regrid function, \code{\link{resample}} for the original resample function
}

