% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncol.R, R/nrow.R, R/resolution.R
\docType{methods}
\name{ncol,scidbst-method}
\alias{ncol,scidbst}
\alias{ncol,scidbst-method}
\alias{nrow,scidbst}
\alias{nrow,scidbst-method}
\alias{res,scidbst}
\alias{resolution,scidbst}
\alias{xres,scidbst-method}
\alias{yres,scidbst-method}
\title{Spatial resolution}
\usage{
\S4method{ncol}{scidbst}(x)

\S4method{nrow}{scidbst}(x)

\S4method{res}{scidbst}(x)

\S4method{xres}{scidbst}(x)

\S4method{yres}{scidbst}(x)
}
\arguments{
\item{x}{scidbst object}
}
\value{
numeric or numeric vector
}
\description{
Returns information about the spatial resolution as numeric values, e.g.(xres,yres), xres or yres or in terms of number of
cells in West-East dimension and South-North dimension (ncol, nrow).
}
\details{
The spatial resolution is calculated with the spatial extent and the number of rows or columns. The number of rows and columns
can be derived by transforming the spatial extent with affine transformation into the dimensional index values and calculating
the differences between the minimum and maximum value for each dimension.
Therefore the spatial resolution describes the distances between the regular parallel lines throughout the image.

In terms of the cell numbers for rows and columns, the array is considered to have 2 dimensions. ncol and nrow refer primarily
to the spatial dimensions. If the array is spatial, then nrow and ncol refer to the number of cells in the
North-South and the West-East axis respectively. If the array is just referenced with a temporal reference system, then nrow
is 1 and ncol is the number of potential units in the temporal extent.

In case both reference systems are present, then nrow and ncol only refer to the spatial dimensions.
}
\examples{
\dontrun{
arr = scidbst("sp_array")
res(arr)
xres(arr)
yres(arr)
nrow(arr)
ncol(arr)
}
}

