% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRS-class.R
\docType{methods}
\name{res,scidbst-method}
\alias{res,scidbst-method}
\alias{xres,scidbst-method}
\alias{yres,scidbst-method}
\title{Spatial resolution}
\usage{
\S4method{res}{scidbst}(x)

\S4method{xres}{scidbst}(x)

\S4method{yres}{scidbst}(x)
}
\arguments{
\item{x}{scidbst}
}
\value{
numeric or numeric vector
}
\description{
Returns informaiton about the spatial resolution as a numeric values, e.g.(xres,yres), xres or yres.
}

