% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb-basic.R
\docType{methods}
\name{scidb_attributes,scidbst-method}
\alias{schema,scidbst}
\alias{scidb_attributes,scidbst}
\alias{scidb_attributes,scidbst-method}
\alias{scidb_coordinate_bounds,scidbst}
\alias{scidb_coordinate_bounds,scidbst-method}
\alias{scidb_coordinate_chunksize,scidbst}
\alias{scidb_coordinate_chunksize,scidbst-method}
\alias{scidb_coordinate_end,scidbst}
\alias{scidb_coordinate_end,scidbst-method}
\alias{scidb_coordinate_overlap,scidbst}
\alias{scidb_coordinate_overlap,scidbst-method}
\alias{scidb_coordinate_start,scidbst}
\alias{scidb_coordinate_start,scidbst-method}
\title{Schema related functions}
\usage{
\S4method{scidb_attributes}{scidbst}(x)

\S4method{scidb_coordinate_bounds}{scidbst}(x)

\S4method{scidb_coordinate_start}{scidbst}(x)

\S4method{scidb_coordinate_end}{scidbst}(x)

\S4method{scidb_coordinate_overlap}{scidbst}(x)

\S4method{scidb_coordinate_chunksize}{scidbst}(x)

\S4method{schema}{scidbst}(x)
}
\arguments{
\item{x}{scidbst array}
}
\value{
list or vector of characters, numerics or a character string
}
\description{
The listed functions are wrapped from the scidb package. The intention of those wrapped functions is the direct access
to strucutural information about the scidb array, e.g. the used schema, overlap, start/end coordinates.
}
\examples{
\dontrun{
 array = scidbst("some_st_array")

 # information operations
 scidb_coordinate_bounds(array) # returns a list
 scidb_attributes(array) # return names of the attributes

 scidb_coordinate_start(array)
 scidb_coordinate_end(array)
 scidb_coordinate_overlap(array)
 scidb_coordinate_chunksize(array)
 # the above return mostly numeric vectors, except the boundary is unbounded with "*" (then a vector of characters)

 schema(array) #returns a string
}
}
\seealso{
\link{dimensions,scidbst}, \link[scidb]{scidb_coordinate_bounds}, \link[scidb]{scidb_attributes}, \link[scidb]{scidb_coordinate_start}
\link[scidb]{scidb_coordinate_end}, \link[scidb]{scidb_coordinate_overlap}, \link[scidb]{scidb_coordinate_chunksize},
\link[scidb]{schema}
}

