% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb-basic.R
\name{scidb_op}
\alias{scidb_op}
\title{Reports the current scidb operation cascade}
\usage{
scidb_op(x)
}
\arguments{
\item{x}{scidb or scidbst object}
}
\value{
character - AFL query
}
\description{
The internal scidb proxy object contains the current AFL query string. scidb_op returns the AFL query string.
}

