% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbeval.R
\docType{methods}
\name{scidbeval,scidbst,character-method}
\alias{scidbeval,scidbst,character-method}
\title{Executes cascaded operations and stores data in scidb under a given name}
\usage{
\S4method{scidbeval}{scidbst,character}(expr, name, eval = TRUE, gc = TRUE,
  temp = FALSE)
}
\arguments{
\item{expr}{The scidbst object}

\item{name}{The name of the target array in which the data is stored}

\item{eval}{A flag whether or not the commands shall be executed in scidb}

\item{gc}{A flag whether or not the result should be tied to the R garbage collector}

\item{temp}{A flag wheter or not the resulting scidb array is temporary}
}
\value{
The modified scidbst object
}
\description{
The function works in the similar way as \link{scidbeval}. As an expression this function requires a 'scidbst' object,
which can be modified by various scidb operations. By calling this function the actual commands are executed in the
SciDB cluster. The result will be stored under the given 'name' parameter. In addition to the original function, the
evaluation of a scidbst object will also set the current spatial and/or temporal reference.
}

