% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbst-class.R
\docType{class}
\name{scidbst-class}
\alias{.scidbst_class}
\alias{scidbst}
\alias{scidbst-class}
\alias{stdb}
\title{Class scidbst}
\usage{
stdb(...)
}
\arguments{
\item{...}{stuff passed to scidb()}
}
\value{
\link{scidbst} object
}
\description{
Class \code{scidbst} inherits from class \code{scidb}

Constructor for scidbst
}
\section{Slots}{

\describe{
\item{\code{CRS}}{The coordinate reference system used as class 'CRS' that represents a Proj.4 string}

\item{\code{extent}}{The outer boundary of the SciDB array in referenced coordinates}

\item{\code{affine}}{The affine transformation used to convert real-world coordinates into image frame coordinates}
}}

