% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{scidbst-package}
\alias{scidbst-package}
\title{Spatio Temporal Extension for SciDBR (scidbst)}
\description{
Spatio Temporal Extension for SciDBR (scidbst)
}
\seealso{
\code{\link{scidbst}}
}

