% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{scidbst-pkg-package}
\alias{scidbst}
\alias{scidbst-package}
\title{Spatio Temporal Extension for SciDB (scidbst)}
\description{
This package extends the package scidb for spatial and temporal referenced dimensions. A set of functions were modified in order
to maintain the correct references, when applying the functions in scidb.
}
\note{
This package requires to SciDB instance to operate on, as well as the scidb package for R. It is also important that the SciDB instance
runs the scidb4geo extension, which allows to maintain the references in scidb.
}
\author{
Florian Lahn <florian.lahn@uni-muenster.de>
}
\seealso{
The scidb4geo extension: \link{https://github.com/mappl/scidb4geo} \\
Docker image for a quick scidb4geo setup: \link{https://github.com/mappl/scidb-eo}
}

