% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scidbst.ls}
\alias{scidbst.ls}
\title{Lists all scidb arrays with a dimension reference}
\usage{
scidbst.ls()
}
\value{
data.frame with columns "name" and "setting"
}
\description{
This function will list all the SciDB arrays that have a special reference on one or more dimensions. It will also list the type of
array: 's' for spatial, 't' for temporal, st for the combination of both
}
\seealso{
\code{\link{scidb::scidbls}}
}

