% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbstconnect.R
\name{scidbstconnect}
\alias{scidbstconnect}
\title{Connect to a SciDB database from R and GDAL}
\usage{
scidbstconnect(host = options("scidb.default_shim_host")[[1]],
  port = options("scidb.default_shim_port")[[1]], username, password,
  auth_type = c("scidb", "digest"), protocol = c("http", "https"))
}
\arguments{
\item{host}{see \code{\link{scidbconnect}}}

\item{port}{see \code{\link{scidbconnect}}}

\item{username}{see \code{\link{scidbconnect}}}

\item{password}{A see \code{\link{scidbconnect}}}

\item{auth_type}{see \code{\link{scidbconnect}}}

\item{protocol}{see \code{\link{scidbconnect}}}
}
\description{
Calls \code{\link{scidbconnect}} and sets aditional environment variables to connect to the database from GDAL.
}
\seealso{
\code{\link{scidbconnect}}
}

