% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbst-class.R
\name{setSelection}
\alias{setSelection}
\title{setSelection}
\usage{
setSelection(x, dim, val)
}
\description{
Sets a selector for the scidbst array. For more information see details.
}
\details{
Usually printing an image needs a reduction into two dimensions. With the selector you can set a dimension to a fixed value, e.g. in a timeseries of spatial coverages you can select the date.
}

