% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRS-class.R
\docType{methods}
\name{setSRS,scidb,SRS,matrix-method}
\alias{setSRS,scidb,SRS,matrix-method}
\title{Sets the spatial reference system to a given array}
\usage{
\S4method{setSRS}{scidb,SRS,matrix}(x, srs, affine, return = FALSE)
}
\arguments{
\item{x}{a scidb array}

\item{srs}{a SRS class}

\item{affine}{a 2x3 matrix containing the affine projection}

\item{return}{logical - if a scidbst object shall be returned (default FALSE)}
}
\value{
a scidbst object if return=TRUE
}
\description{
By stating the srs and the affine projection together with a given scidb array, that array
becomes a spatial array.
}

