% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRS-class.R
\docType{methods}
\name{setSRS,scidb}
\alias{setSRS}
\alias{setSRS,scidb}
\alias{setSRS,scidbst}
\title{Sets the spatial reference system at a given array}
\usage{
\S4method{setSRS}{scidb,SRS,matrix}(x, srs, affine, return = FALSE)
}
\arguments{
\item{x}{a \link{scidb} array}

\item{srs}{a \link{SRS} class}

\item{affine}{a 2x3 matrix containing the affine projection}

\item{return}{logical - if a scidbst object shall be returned (default FALSE)}
}
\value{
a scidbst object if return=TRUE
}
\description{
By knowing the \code{\link{SRS}} and the \code{\link{affine}} projection a spatial reference can be established on a given
scidb array.
}

