% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRS-class.R
\docType{methods}
\name{setTRS,scidb,TRS-method}
\alias{setTRS,scidb,TRS-method}
\title{Add a TRS to a scidb array}
\usage{
\S4method{setTRS}{scidb,TRS}(x, trs, return = FALSE)
}
\arguments{
\item{x}{scidb}

\item{trs}{TRS}

\item{return}{logical - if a scidbst object shall be returned (default FALSE)}
}
\value{
scidbst if parameter return=TRUE
}
\description{
The function adds a temporal reference to a scidb array and returns a scidbst object
}
\note{
Using \code{setTRS} will also persist the metadata for the array directly in SciDB.
}

