% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbst-class.R
\docType{methods}
\name{show,scidbst-method}
\alias{show,scidbst-method}
\title{Show scidbst object}
\usage{
\S4method{show}{scidbst}(object)
}
\arguments{
\item{object}{scidbst object}
}
\description{
Creates a printable representation about relevant information about the scidbst object.
}

