% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemporalExtent-class.R
\docType{methods}
\name{show,TemporalExtent-method}
\alias{show,TemporalExtent-method}
\title{Show method}
\usage{
\S4method{show}{TemporalExtent}(object)
}
\arguments{
\item{x}{TemporalExtent}
}
\description{
creates a print version of the TemporalExtent object
}

