% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbst-class.R
\docType{methods}
\name{slice,scidbst,character,numeric-method}
\alias{slice,scidbst,character,numeric-method}
\title{Slice the array}
\usage{
\S4method{slice}{scidbst,character,numeric}(x, d, n)
}
\arguments{
\item{x}{\code{scidb} array object}

\item{d}{name or numeric position of dimension to slice}

\item{n}{coordinate value to slice on}
}
\value{
scidbst A new scidbst object with reduced number of dimensions
}
\description{
Takes a dimension name and a value to create a slice of an array. This usually means reducing the dimensions
of an array.
}

