% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\docType{methods}
\name{slice,scidbst}
\alias{slice,scidbst}
\alias{slice,scidbst,character,POSIXt-method}
\alias{slice,scidbst,character,integer-method}
\alias{slice,scidbst,character,numeric-method}
\alias{slice,scidbst,missing,POSIXt-method}
\title{Slice a scidbst object at a particular dimension value}
\usage{
\S4method{slice}{scidbst,character,character}(x, d, n)

\S4method{slice}{scidbst,character,numeric}(x, d, n)

\S4method{slice}{scidbst,character,integer}(x, d, n)

\S4method{slice}{scidbst,missing,POSIXt}(x, d, n)

\S4method{slice}{scidbst,character,POSIXt}(x, d, n)
}
\arguments{
\item{x}{scidbst array object}

\item{d}{name of a dimension}

\item{n}{coordinate value to slice on, e.g. a POSIXt, POSIXt conform date/time string, temporal index}
}
\value{
scidbst array with a reduced number of dimensions.
}
\description{
Takes a dimension name and a value to create a slice of an array. This usually means reducing the dimensions
of an array by one. The dimension name and value choosen fixed and the other dimensions and attributes are returned.
}
\examples{
\dontrun{
scidbconnect(...)
scidbst.obj = scidbst(array_name) #scidb array with spatial and temporal dimension

#examples on the temporal dimension
slice1 = slice(scidbst.obj,"t","0") # @ temporal index 0
slice2 = slice(scidbst.obj,"t",1) # @ temporal index 1
slice3 = slice(scidbst.obj,"t","2016-05-05") # @ the temporal index for a date
# other existing dimensions are also applicable
}
}
\seealso{
\link[scidb]{slice}
}

