% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spplot.R
\docType{methods}
\name{spplot,scidbst-method}
\alias{spplot,scidbst-method}
\title{spplot for scidbst objects}
\usage{
\S4method{spplot}{scidbst}(obj, ...)
}
\arguments{
\item{obj}{scidbst object}

\item{...}{Any argument that can be passed to spplot and levelplot}

\item{maxpixels}{Integer. Number of pixels to sample from each layer of large Raster objects}
}
\description{
Like the raster version for spplot, this function plots a spatially referenced scidb array. To make this work on a
multidimensional array, the number of dimensions must be reduced to two spatial dimensions. This can
be done by using the 'slice' operation of scidb.
}
\seealso{
\code{\link[raster]{spplot,Raster-method}}
}

