% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRS-class.R
\docType{methods}
\name{srs,scidbst-method}
\alias{srs,scidbst-method}
\title{Returns the spatial reference}
\usage{
\S4method{srs}{scidbst}(x)
}
\arguments{
\item{x}{scidbst}
}
\value{
\code{\link{SRS}} class
}
\description{
The function returns the SRS object of a scidbst object.
}

