% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\docType{methods}
\name{stream,scidbst,function-method}
\alias{stream,scidbst}
\alias{stream,scidbst,function-method}
\title{Apply a R script on all chunks of an array}
\usage{
\S4method{stream}{scidbst,`function`}(x, f, array, packages = c(),
  parallel = FALSE, cores = 1, aggregates = c(), output, logfile = NULL,
  ...)
}
\arguments{
\item{x}{scidbst array}

\item{array}{the array name to store the results under}

\item{packages}{the names of the R packages that are used during the function call}

\item{parallel}{property in ddply to perform parallel processing}

\item{cores}{number of cores used for parallel processing on a single instance}

\item{aggregates}{the attribute names which are used to group by}

\item{output}{a list with the name and the data type in SciDB as key-value pairs}

\item{logfile}{the file path used to log during the processing}
}
\description{
This function will execute a R function within a ddply call to perform higher level analysis using the [streaming interface of SciDB](https://github.com/Paradigm4/stream).
}
\seealso{
\link[scidbst]{r.apply}
}

