% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\docType{methods}
\name{subset.scidbst}
\alias{subset,scidbst-method}
\alias{subset.scidbst}
\title{Subset method for scidbst objects}
\usage{
\method{subset}{scidbst}(x, ...)

\S4method{subset}{scidbst}(x, ...)
}
\arguments{
\item{x}{scidbst object}

\item{...}{filter expression}
}
\value{
a scidbst object
}
\description{
This function wraps scidbs subset function for scidbst objects. Internally, scidb uses its \code{filter} function to create the
subset. To formulate the filter expression, please have a look at the scidb documentation (\code{\link{subset.scidb}}).
Since scidbs filter function is used, the resulting array will probably be sparse.
}
\note{
It is recommended to use the string based SciDB filter syntax.
}
\examples{
\dontrun{
scidbconnect()
ls7_ndvi = scidbst("LS7_BRAZIL_NDVI")
ls7_ndvi_gr05 = subset(ls7_ndvi, "NDVI > 0.5")
}
}
\seealso{
\code{\link{subset.scidb}}
}

