% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textent.R
\docType{methods}
\name{t.extent,scidbst-method}
\alias{t.extent,scidbst-method}
\title{Extracts the temporal extent of an scidbst array}
\usage{
\S4method{t.extent}{scidbst}(x)
}
\arguments{
\item{x}{a scidbst object}
}
\value{
The temporal extent as a \link{\code{TemporalExtent}} object
}
\description{
This method will return the stored temporal extent that contains the minimum and maximum date values in a list. The list will be
empty if the array has no temporal dimension.
}

