% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRS-class.R
\docType{methods}
\name{t0}
\alias{t0}
\alias{t0,scidbst-method}
\title{Returns the temporal datum (t0)}
\usage{
\S4method{t0}{TRS}(x)

\S4method{t0}{scidbst}(x)
}
\arguments{
\item{x}{TRS object}

\item{x}{scidbst object}
}
\value{
POSIXt value
}
\description{
This function returns the reference t0 value. In combination with scidbst this refers to the 0 value in the temporal dimension.
}

