% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRS-class.R
\docType{methods}
\name{tdim}
\alias{tdim}
\alias{tdim,scidbst-method}
\title{Get temporal dimension name}
\usage{
\S4method{tdim}{TRS}(x)

\S4method{tdim}{scidbst}(x)
}
\arguments{
\item{x}{scidbst object}
}
\value{
The temporal dimension name or the first dimension name
}
\description{
Returns the name of the temporal dimension of a scidbst array. If the object is not temporal it will
return the first dimension name.
}

