% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemporalExtent-class-decl.R, R/TemporalExtent-class.R
\docType{class}
\name{TemporalExtent}
\alias{.textent_class}
\alias{TemporalExtent}
\alias{textent}
\title{Temporal Extent class}
\usage{
textent(min, max)
}
\arguments{
\item{min}{A POSIXt derived object for the minimum inclusive boundary}

\item{max}{A POSIXt derived object for the maximum inclusive boundary}
}
\value{
TemporalExtent object
}
\description{
This class represents a temporal interval with POSIXt dates.

This constructor function creates a TemporalExtent object that defines an interval on the temporal dimension.
}
\section{Slots}{

\describe{
\item{\code{min}}{POSIXt derived object defining the minimum inclusive boundary}

\item{\code{max}}{POSIXt derived object defining the maximum inclusive boundary}
}}
\examples{
tmin = as.POSIXlt("2016-01-01")
tmax = strptime("01-02-2016",format="\%d-\%m-\%Y")
ext = textent(tmin,tmax)
}

