% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{tindex,scidbst-method}
\alias{tindex,scidbst-method}
\title{Calculates the temporal index of a date time string or POSIXt object}
\usage{
\S4method{tindex}{scidbst}(x, time)
}
\arguments{
\item{x}{scidbst array}

\item{time}{a character string representing a valid date time or a POSIXt object}
}
\value{
integer the nearest temporal index for specified date time
}
\description{
Calculates the temporal index of a date time string or POSIXt object
}

