% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemporalExtent-class.R
\docType{methods}
\name{tmax}
\alias{tmax}
\alias{tmax,scidbst-method}
\title{Maximum temporal boundary}
\usage{
\S4method{tmax}{TemporalExtent}(x)

\S4method{tmax}{scidbst}(x)
}
\arguments{
\item{x}{object}
}
\value{
POSIXt object referring to the maximum boundary
}
\description{
Return the maximum boundary of a TemporalExtent object
}

