% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemporalExtent-class.R
\docType{methods}
\name{tmin}
\alias{tmin}
\alias{tmin,scidbst-method}
\title{Minimum temporal boundary}
\usage{
\S4method{tmin}{TemporalExtent}(x)

\S4method{tmin}{scidbst}(x)
}
\arguments{
\item{x}{object}
}
\value{
POSIXt object referring to the minimum boundary
}
\description{
Return the minimum boundary of a TemporalExtent object
}

