% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-apply.R
\docType{methods}
\name{transform.scidbst}
\alias{apply.scidbst}
\alias{transform,scidbst-method}
\alias{transform.scidbst}
\title{Transform / Apply function for scidbst object}
\usage{
\method{transform}{scidbst}(`_data`, ...)

\S4method{transform}{scidbst}(`_data`, ...)
}
\arguments{
\item{_data}{The object to be transformed}

\item{...}{Further arguments of the form \code{tag=value}}
}
\description{
This function overrides the S3method 'transform' and provides the same functionality as the 'apply' in SciDB.
}
\seealso{
\link{transform}
}

