% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transformAllSpatialIndices}
\alias{transformAllSpatialIndices}
\title{Transform all spatial indices to spatial coordinates}
\usage{
transformAllSpatialIndices(obj, df)
}
\arguments{
\item{obj}{The scidbst object}

\item{df}{A data.frame derived from an scidbst object}
}
\value{
data.frame with spatial coordinates
}
\description{
Transform all spatial indices to spatial coordinates
}

