% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRS-class.R
\docType{methods}
\name{tres}
\alias{tres}
\alias{tres,scidbst-method}
\title{Returns the temporal resolution}
\usage{
\S4method{tres}{TRS}(x)

\S4method{tres}{scidbst}(x)
}
\arguments{
\item{x}{TRS object}

\item{x}{scidbst object}
}
\value{
numeric value
}
\description{
This function returns the temporal resolution as a numeric value with the unit stated by 'tunit'.
}

