% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRS-class.R
\docType{methods}
\name{trs,scidbst-method}
\alias{trs,scidbst-method}
\title{Returns the Temporal reference object}
\usage{
\S4method{trs}{scidbst}(x)
}
\arguments{
\item{x}{scidbst object}
}
\value{
\code{\link{TRS} object}
}
\description{
Returns the Temporal reference object
}

