% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRS-class.R
\docType{methods}
\name{tunit}
\alias{tunit}
\alias{tunit,scidbst-method}
\title{Returns the temporal unit}
\usage{
\S4method{tunit}{TRS}(x)

\S4method{tunit}{scidbst}(x)
}
\arguments{
\item{x}{TRS object}

\item{x}{scidbst object}
}
\value{
character value
}
\description{
This function returns the temporal unit as a string (character)
}

