% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRS-class.R
\docType{methods}
\name{xdim}
\alias{xdim}
\alias{xdim,SRS-method}
\title{Get spatial x dimension name}
\usage{
\S4method{xdim}{scidbst}(x)

\S4method{xdim}{SRS}(x)
}
\arguments{
\item{x}{scidbst object}
}
\value{
The name of the spatia West-East dimension or the first dimension name
}
\description{
Returns the name of the spatial dimension of a scidbst array, that relates to a West-East dimension. If the object is not
spatially referenced it will return the second dimension name.
}

