% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgrid.R
\docType{methods}
\name{xgrid,scidbst-method}
\alias{xgrid,scidbst-method}
\title{Enlarge function}
\usage{
\S4method{xgrid}{scidbst}(x, grid, expr = "S")
}
\arguments{
\item{x}{scidbst array}

\item{grid}{vector of integers which correspond to each dimension or a scidbst array}

\item{expr}{character specifiying the type (one of c("S","T","ST")) that determins the dimensions to use for scaling}
}
\value{
a modified version of x (scidbst array) with changed dimensionality
}
\description{
This function can be used to enlarge the dimensional resolution of an array, e.g. changing the spatial resolution
from a low resolution to a higher one. It calls the same called function in SciDB to execute the operation. There
will be no interpolation of new values. The "old" cell value will be repeated the amount of times stated with the
'grid' statement.
}
\note{
It is important to mention, that the dimensional scaling factor needs to be a integer value, other values will
be rounded to the nearest integer. This is due to the fact that SciDB operates with int64 values as dimensional values.
}

