% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\docType{methods}
\name{aggregate.sp,scidbst-method}
\alias{aggregate.sp,scidbst-method}
\title{aggregates over space}
\usage{
\S4method{aggregate.sp}{scidbst}(x, by, ...)
}
\value{
scidbst object with the spatial resolution extended to the whole image extent
}
\description{
This function aggregates over space leaving a scidbst array that has a spatial resolution of the
prior spatial extent of the array.
}

