% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbst-class.R
\docType{methods}
\name{aggregate.t,scidbst-method}
\alias{aggregate.t,scidbst-method}
\title{aggregates over time}
\usage{
\S4method{aggregate.t}{scidbst}(x, attributes, ...)
}
\description{
aggregates over time
}

