% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\docType{methods}
\name{aggregate.t,scidbst-method}
\alias{aggregate.t}
\alias{aggregate.t,scidbst-method}
\title{Aggregate over time}
\usage{
aggregate(x, by, FUN, window, variable_window)
}
\arguments{
\item{x}{A \code{scidbst} object.}

\item{by}{optional single character string or a list of array dimension and/or attribute names to group by additionally to
the temporal dimension}

\item{FUN}{a character string representing a SciDB aggregation expression or a reduction function.}

\item{window}{optional, if specified, perform a moving window aggregate along the specified coordinate windows.}

\item{variable_window}{optional, if specified, perform a moving window aggregate over successive data values along the
coordinate dimension axis specified by \code{by}.}
}
\value{
scidbst array with aggregated values on the temporal dimension (no temporal dimension)
}
\description{
This function aggregates a scidbst array for the temporal dimension. This means the resulting array will stripped from
the temporal dimension and the values on the temporal dimension will be aggregated.
}
\note{
The aggregated scidbst object will loose its temporal reference after aggregation, due to the fact that the spatial
dimensions are removed
}
\examples{
\dontrun{
scidbconnect(...)
scidbst.obj = scidbst(array_name)
aggt = aggregate.t(scidbst.obj,FUN="avg(attribute1)") # returns something similar to a raster with aggregated values over the temporal dimension
}
}

