% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRS-class.R
\docType{methods}
\name{copyTRS,scidbst,scidbst-method}
\alias{copyTRS,scidb,scidbst-method}
\alias{copyTRS,scidbst,scidbst-method}
\title{Copy temporal reference}
\usage{
\S4method{copyTRS}{scidbst,scidbst}(x, y)

\S4method{copyTRS}{scidb,scidbst}(x, y)
}
\arguments{
\item{x}{scidbst or scidb object}

\item{y}{scidbst object}
}
\value{
modified x
}
\description{
Copies the temporal reference systems from scidbst object y to scidbst object x.
}
\examples{
\dontrun{
 st.arr1 = scidbst("st_arr_1")
 st.arr2  = scidbst("st_arr_2")
 simple.arr = scidb("some_array")

 # overwrite trs of array 1 with trs of array 2
 st.arr1 = copyTRS(st.arr1, st.arr2)

 # set the TRS of array 2 for the simple array (having the same named temporal dimension)
 t.arr1 = copyTRS(simple.arr, st.arr2)
}
}

