% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\docType{methods}
\name{crs,scidbst}
\alias{crs,scidbst}
\title{Gets the Coordinate Reference System}
\usage{
\S4method{crs}{scidbst}(x, ...)
}
\arguments{
\item{x}{scidbst object}
}
\value{
CRS object

\link[rgdal]{CRS}
}
\description{
Returns the coordinate reference system of a scidbst object in form of a CRS object.
}

