% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent.R
\docType{methods}
\name{extent,scidbst-method}
\alias{extent,scidbst-method}
\alias{extent<-,ANY,Extent-method}
\alias{extent<-,scidbst,Extent-method}
\title{Gets and sets the spatial extent of a scidb/scidbst array}
\usage{
\S4method{extent}{scidbst}(x)

\S4method{extent}{scidbst,Extent}(x) <- value

\S4method{extent}{ANY,Extent}(x) <- value
}
\description{
This function let you retrieve the spatial extent as a Extent-class object or it
assigns or replaces an extent for a scidb/scidbst object.
}

