% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbst-class.R
\docType{methods}
\name{readAll,scidbst-method}
\alias{readAll,scidbst-method}
\title{readAll}
\usage{
\S4method{readAll}{scidbst}(object)
}
\description{
In combination with \code{raster::getValues()} this function is called to retrieve the values from the source and store them in memory
}

