% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbst-class.R
\docType{methods}
\name{readAll,scidbst-method}
\alias{readAll,scidbst-method}
\title{readAll}
\usage{
\S4method{readAll}{scidbst}(object)
}
\arguments{
\item{object}{scidbst object}
}
\value{
the modified scidbst object with values
}
\description{
Like \code{raster::getValues()} this function is called to retrieve the values from the source and store them in memory. It
differs from getValues in the fact that the scidbst object will be manipulated and returned back.
}

