% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbst-class-decl.R, R/scidbst-class.R
\docType{class}
\name{scidbst}
\alias{scidbst}
\title{Class scidbst}
\usage{
scidbst(...)
}
\arguments{
\item{...}{parameter that are passed on to \code{\link[scidb]{scidb}}}
}
\value{
scidbst object
}
\description{
The scidbst class contains several objects in order to describe a SciDB array spatially and temporally. Therefore the
object holds a reference to the SciDB array via a scidb object and the dimension references for space and time with a
SRS and TRS object respectively.

Constructor for scidbst
}
\section{Slots}{

\describe{
\item{\code{extent}}{The outer boundary of the SciDB array in referenced coordinates}

\item{\code{affine}}{The affine transformation used to convert real-world coordinates into image frame coordinates}

\item{\code{tExtent}}{the temporal extent as a \code{\link{TemporalExtent}} object}

\item{\code{isSpatial}}{A flag whether or not this object has a spatial reference}

\item{\code{isTemporal}}{A flag whether or not this object has a temporal reference}

\item{\code{trs}}{The temporal reference object (\code{\link{TRS}})}

\item{\code{srs}}{The spatial reference object \code{\link{SRS}}}

\item{\code{proxy}}{scidb proxy \code{\link[scidb]{scidb}}}

\item{\code{extent}}{spatial \code{\link[raster]{Extent}}}

\item{\code{title}}{The name of the original array}
}}
\note{
At least parameter \code{name} should be provided
}
\examples{
\dontrun{
scidbconnect(host,port,user,pwd,protocol,authtype)
chicago = scidbst(name="chicago_sts")
}
}

