% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbst-class.R
\docType{methods}
\name{spplot,scidbst-method}
\alias{spplot,scidbst-method}
\title{spplot for scidbst objects}
\usage{
\S4method{spplot}{scidbst}(obj, maxpixels = 50000, as.table = TRUE, zlim,
  ...)
}
\arguments{
\item{obj}{scidbst object}
}
\description{
Like the raster version for spplot, this function plots a spatially referenced scidb array. To make this work on a
multidimensional array, the number of dimensions must be reduced to the correct two spatial dimensions. This can
be done by using the 'slice' operation of scidb.
}

