% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spplot.R
\docType{methods}
\name{spplot,scidbst}
\alias{spplot,scidbst}
\title{spplot for scidbst objects}
\usage{
\S4method{spplot}{scidbst}(obj, ...)
}
\arguments{
\item{obj}{scidbst object}

\item{...}{Any argument that can be passed to spplot and levelplot}

\item{maxpixels}{Integer. Number of pixels to sample from each layer of large Raster objects}
}
\description{
Like the raster version for spplot, this function plots a spatially referenced scidb array. To make this work on a
multidimensional array, the number of dimensions must be reduced to two spatial dimensions. This can
be done by using the 'slice' operation of scidb. To plot the array, the data is first downloaded to your machine and then
transformed to a RasterBrick object to be plotted.
}
\details{
The spplot will first assume that the array is spatially referenced and will use the North-South and West-East dimensions to
match the rows and columns for the raster brick. If the array is not spatially referenced, but temporal the spplot will
result a 1 row x t columns (t = number of time intervals) image.
}
\seealso{
\code{\link[raster]{spplot}}
}

