% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRS-class-decl.R, R/SRS-class.R
\docType{class}
\name{SRS}
\alias{.srs_class}
\alias{SRS}
\title{Spatial Reference Class}
\usage{
SRS(projargs, dimnames)
}
\arguments{
\item{projargs}{character - The projection arguments as defined by PROJ4}

\item{dimnames}{character - The dimension names}
}
\description{
This class holds information about the spatial dimensions. In particular, the dimension names and the projection arguments. The projection
arguments are inherited from the \code{\link[sp]{CRS}} class

Creates a new SRS object with projection arguments and dimension names (y,x)-order
}
\section{Slots}{

\describe{
\item{\code{dimnames}}{character - The names for the dimensions in the order y-axis (South-North), x-axis(West-East)}

\item{\code{authority}}{character - The name of the authority that provides the information about the reference system}

\item{\code{srid}}{integer - The integer identifier that 'names' the spatial reference system}

\item{\code{srtext}}{character - The textual description of the reference system in OGC Well Known Text (WKT) format}
}}

