% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemporalExtent-class-decl.R
\docType{class}
\name{TemporalExtent}
\alias{.textent_class}
\alias{TemporalExtent}
\title{Temporal Extent class}
\description{
This class represents a temporal interval with POSIXt dates. The constructor for this class is the function
\code{\link{textent,scidbst}}.
}
\section{Slots}{

\describe{
\item{\code{min}}{POSIXt derived object defining the minimum inclusive boundary}

\item{\code{max}}{POSIXt derived object defining the maximum inclusive boundary}
}}
\examples{
tmin = as.POSIXlt("2016-01-01")
tmax = strptime("01-02-2016",format="\%d-\%m-\%Y")
ext = textent(tmin,tmax)

}
\seealso{
\code{\link{textent,scidbst}}
}

