% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-apply.R
\docType{methods}
\name{transform.scidbst}
\alias{apply.scidbst}
\alias{transform,scidbst-method}
\alias{transform.scidbst}
\title{Transform / Apply function for scidbst object}
\usage{
\method{transform}{scidbst}(`_data`, ...)

\S4method{transform}{scidbst}(`_data`, ...)
}
\arguments{
\item{_data}{The object to be transformed}

\item{...}{Further arguments of the form \code{tag=value}}
}
\description{
This function applies arithmetic operations on the attributes of an SciDB array.
}
\examples{
\dontrun{
ls7_brazil_regrid = scidbst("LS7_brazil_regridded")
# Calculating the NDVI and MDVI of a Landsat 7 scene
ls7_calc = transform(ls7_brazil_regrid, ndvi = "(band4 - band3) / (band4 + band3)", mdvi = "(band8 - band3) / (band8 + band3)")
}

}
\seealso{
\link{transform}, \link{transform.scidb}
}

