% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transformAllSpatialIndices}
\alias{transformAllSpatialIndices}
\title{Transform all spatial indices to spatial coordinates}
\usage{
transformAllSpatialIndices(obj, df)
}
\arguments{
\item{obj}{The scidbst object}

\item{df}{A data.frame derived from an scidbst object with indices as dimension values}
}
\value{
data.frame with spatial coordinates
}
\description{
Transform all spatial indices to spatial coordinates
}
\examples{
\dontrun{
 x = scidbst(arrayname)
 df = as(x,"data.frame") # executes iquery and returns a data.frame with dimension indices and attribute values
 coordinates = transformAllSpatialIndices(x,df) # returns only the spatial coordinates of scidbst object
}
}

