% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transformAllTemporalIndices}
\alias{transformAllTemporalIndices}
\title{Transforms a given data.frame with temporal dimension into dates by a given scidbst object}
\usage{
transformAllTemporalIndices(obj, df)
}
\arguments{
\item{obj}{The scidbst object}

\item{df}{A data.frame derived from an scidbst object}
}
\value{
data.frame with temporal coordinates
}
\description{
This function transforms the temporal indices of a data.frame that was derived from a scidbst object into the actual
dates.
}
\examples{
\dontrun{
 x = scidbst(arrayname)
 df = as(x,"data.frame") # executes iquery and returns a data.frame with dimension indices and attribute values
 times = transformAllTemporalIndices(x,df) # returns only the unique temporal coordinates of scidbst object
}
}

